APP-9414 Export Triangle to create AnnouncementBlock in ACE
