+++ src/Triangle/index.ts
@@ -1,0 +1,5 @@
+import Triangle from './Triangle'
+
+export { default as TriangleProps } from './Triangle.interface'
+
+export default Triangle
